refactor(weekly-report): extract monolith into sbir_etl/reporting/weekly
Stages 1-5 of specs/weekly-awards-report-refactor, protected end-to-end
by the stage-0 golden-file test (byte-identical normalized output before
and after).

- scripts/data/weekly_awards_report.py: 2,807 -> 116 lines; argparse +
  WeeklyAwardsReportBuilder(...).run(). CLI flags unchanged (weekly.yml
  only consumes --output).
- New package sbir_etl/reporting/weekly/: models (CompanyResearch,
  SolicitationTopic), fetching (CSV resolution, DuckDB window query,
  clean/dedup, history wrappers), enrichment (rate limiters, stage
  budgets, all lookup_*/fetch_* wrappers, inflation/district/BEA),
  llm_digests (pure prompt-context builders), llm (OpenAI stages),
  rendering (pure markdown), link_verification, debug (shared --debug
  switch replacing the module global), orchestrator (pipeline port of
  the old main()).
- Function bodies moved verbatim (AST-cut blocks); only mechanical
  changes: package imports replacing in-file references, debug_enabled()
  replacing the DEBUG global, and three latent typing fixes surfaced by
  mypy now checking this code (float accumulators, str|None key).
- Dropped the script's unused local FederalAwardSummary dataclass
  (shadowed the imported alias; zero references).
- New tests: per-function rendering tests, fetching clean/dedup tests
  (28 total in tests/unit/reporting/). mypy sbir_etl clean (189 files).
- Spec tasks.md updated with per-task evidence; T2.3, half of T3.2,
  T5.2 targets, and T5.3 remain open (annotated inline).
