+++ sbir_etl/reporting/weekly/__init__.py
@@ -1,0 +1,6 @@
+"""Weekly SBIR awards report pipeline.
+
+Extracted from scripts/data/weekly_awards_report.py (spec:
+weekly-awards-report-refactor). The CLI script remains the entry point;
+it drives :class:`sbir_etl.reporting.weekly.orchestrator.WeeklyAwardsReportBuilder`.
+"""
